order of tags matters
it turns out that the XML gets processed sequentially, I guess.
So, for example, I had `get` of a class also call `get` on its components (attributes, the list of parents, operations, slots). This meant that the tag for adding a class will be outputted and then directly followed by the components and the tags `changeParent` that belong to that class before moving on to the next class.
Then XModeler will read the file and create the class and attempt to "let it inherit" from classes that do not exist yet.
When I tried opening the file in XModeler it showed an error message along the lines of "did not find new parent at ... ".
So, I changed the order in the XML so that all classes are declared and only then the tags `changeParent` occurred. That fixed the problem.
So, what I did in this commit is to just to let it output everything in the exact same order as in an XML generated by XModeler.