changed `valid` and `Valid`
included a field that refers to the containing class in each of the following:
Operation, Slot, Attribute, in order to eliminate the need for a second parameter for context for checking validity.
To be honest, I find checking the validity like this not very useful because eventually, I will be generating the MLMs anyways. I might resort to making a function that returns a valid "space" for things that could be inserted/changed in an element, instead.
But for now, I will keep it as it is until I figure out other things.
Also, I think it makes sense to leave an option to create an invalid MLM on purpose.
I hope I am not biting more than I can chew.