instantiated `Show` for data types
this is just a proposal to solve the issue "duplication of data".
I had asked you before about this, but these seemed to be misunderstanding.
So, now, if the issue was just that when an MLM is printed on screen, it was absurd and full of duplication, then that issue is solved.
If the issue is along the lines of
"You should not be storing, for example, a copy of the whole class in which an attribute lies, inside that attribute, but instead just refer to it by its name. (store its name as a String or something like that)."
, then the issue is not solved yet, and I would have to re-do the types and navigate MLMs using `find` and `lookup` or something like that.