Smaller contexts for validation
no more passing the whole MLM as an argument to validate an element of that MLM.
Instead, we pass more precise contexts as arguments.
I tried to delegate as many checks as possible to the instances of `Validatable` of inner elements in hope of having a somewhat hierarchical structure.
But to be honest, I am not satisfied with this.
Just the amount of renaming arguments makes the code hard to read.

I might "flatten" even more and just give up on using a class `Validatable` and instead, do most checks at the data type `MLM` itself.