checking for attributes adjusted
Now I check whether a class "has a purpose" which means that it is already linked using some Links or can instantiate an attribute or instantiate and operation (? if implemented in XModeler in the future, or it exists already, but I have not noticed it) or is just a metaclass and thus does not have to do any of the previous.
But also added checking whether an attribute that can be instantiated in a class and has a lower bound of multiplicity larger than 0 is actually instantiated or not.