fixed the test for adding links
It fails sometimes. I am investigating...
Example of this failure is the MLM in the file testingAddLinks.xml
The two links in testingAddLinks_theseCanBeAdded.txt , as the name states, can be added to that MLM, and my function `addLinks` seems to not add them. Or in other words, it adds n-2 links instead of n possible ones.