renamed some parameters that start with `chance`
to `tendency ... `
To express that this is not a guaranteed number and it does not make sense to be treated as a chance for something to actually happen. It is rather just a chance to attempt to do something but only when it is allowed. (Tendency to do ... when possible)